+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## sampquery | Changelog tracking
 
+### v0.0.5 | 10.05.2025 [0f57685](https://github.com/larayavrs/sampquery/commit/0f5768517c83843ae2159e585d7573ad8416ebcf)
+
+- Now [you can get the lagcomp mode](./examples/new-example.py) using the **`lagcomp()`** method.
+- We now support player limiter queries using exception handler.
+
 ### v0.0.4 | 10.05.2025 [e6a6a0](https://github.com/larayavrs/sampquery-dev/commit/e6a6a046f3a3a38d61d8209ce6c2b561e5743281)
 
 - You can now get the detailed player list data using the **`detailed_players()`** method.